TWITTER SENTIMENTAL ANALYSIS
This project explain the tweet whether the tweet is positive or negative and it show the pedection values where we can use TF-IDF model  and logistic regression and we remove the null values and hash tags.The Web is a huge virtual space where to express and share individual opinions, influencing for marketing and communication alike. Reviews and ratings on the internet are increasing their importance in the evaluation of the products and services by potential customers. Social media platforms like Twitter, Facebook, YouTube generate huge amounts of data that can be mined in various ways to understand trends, public sentiments and opinions. Social media data today has become relevant for branding, marketing and business as a whole. A sentiment analyser learns about various sentiments behind a "content piece" through machine learning and predicts the same using AI. Twitter data is considered as a definitive entry point to practice sentimental analysis machine learning problems. Through this data one can get captivating blend of tweet contents and the other related metadata such as hashtags, retweets, locations and much more which pave a way for insightful analysis. Monitoring the social media activities is a good way to measure customer's loyalty, keeping a track on their sentiments towards brands or products. The foremost problem is to build a model to classify tweets and positive, negative or neutral.